Rozdelit funkce podle funkcionalit
